requirements_ocr.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
baidu-aip>=4.16.0
flash-attn>=2.0.0
accelerate>=0.20.0
safetensors>=0.3.0
